pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-regression-detector"
version = "1.0.0"
description = "CI/CD pipeline for detecting LLM output quality regressions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "jinja2>=3.1",
    "requests>=2.31",
    "pandas>=2.0",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.30"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
model-eval = "run_eval:main"

[tool.setuptools.packages.find]
include = ["src*"]
